+++ UE4/UnLua/ELuaProfiler/Source/Public/ELuaTraceInfoNode.h
@@ -30,25 +30,25 @@ struct ELUAPROFILER_API FELuaTraceInfoNode
 	FString Name = "anonymous";
 
 	/* call time */
-	int64 CallTime = 0;
+	double CallTime = 0;
 
 	/* self time */
-	int64 SelfTime = 0;
+	double SelfTime = 0;
 
 	/* total time */
-	int64 TotalTime = 0;
+	double TotalTime = 0;
 
 	/* the total size of lua_State when this node invoke */
-	int32 CallSize = 0;
+	float CallSize = 0;
 
 	/* the total size of lua_State when this node return */
 	//int32 ReturnSize;
 
 	/* the size of this node alloc */
-	int32 AllocSize = 0;
+	float AllocSize = 0;
 
 	/* the size of this node release */
-	int32 GCSize = 0;
+	float GCSize = 0;
 
 	/* the num of calls */
 	int32 Count = 0;
